move crowding to its own module
